Move Log from util to logger module
