+++ src/common/profile.cpp
@@ -8,6 +8,8 @@
 #include <stream/datastream.h>
 #include <util.h>
 
+#include "logger.h"
+
 using namespace std;
 using namespace xengine;
 
@@ -41,7 +43,7 @@ bool CProfile::Save(std::vector<unsigned char>& vchProfile)
 
         encoder.Encode(vchProfile);
     }
-    catch (exception& e)
+    catch (const exception& e)
     {
         StdError(__PRETTY_FUNCTION__, e.what());
         return false;
@@ -115,7 +117,7 @@ bool CProfile::Load(const vector<unsigned char>& vchProfile)
             is >> destOwner.prefix >> destOwner.data;
         }
     }
-    catch (exception& e)
+    catch (const exception& e)
     {
         StdError(__PRETTY_FUNCTION__, e.what());
         return false;
