+++ src/mpvss/parallel.h
@@ -18,7 +18,7 @@
 #include <type_traits>
 #include <vector>
 
-#include "util.h"
+#include "logger.h"
 
 /**
  * Parallel computer for CPU intensive computing
@@ -69,7 +69,7 @@ public:
                     }
                     return true;
                 }
-                catch (std::exception& e)
+                catch (const std::exception& e)
                 {
                     xengine::StdError(__PRETTY_FUNCTION__, e.what());
                     return false;
