+++ src/xengine/CMakeLists.txt
@@ -10,6 +10,7 @@ set(sources
     version.h
     type.h
     util.cpp                util.h
+    logger.cpp              logger.h
     rwlock.h
     cache.h
     compacttv.h
