+++ src/xengine/docker/config.cpp
@@ -12,7 +12,7 @@
 #include <iostream>
 #include <sstream>
 
-#include "util.h"
+#include "logger.h"
 
 using namespace std;
